+++ bootstrap.sh
@@ -5,6 +5,7 @@ if [ -x "$(which python)" ]; then
 
     virtualenv --clear --no-site-packages -p python py2
 
+    ./py2/bin/pip install wheel
     ./py2/bin/pip install coverage
     ./py2/bin/pip install zope.interface
     ./py2/bin/pip install zope.lifecycleevent
@@ -19,6 +20,7 @@ if [ -x "$(which python3)" ]; then
 
     virtualenv --clear --no-site-packages -p python3 py3
 
+    ./py3/bin/pip install wheel
     ./py3/bin/pip install coverage
     ./py3/bin/pip install zope.interface
     ./py3/bin/pip install zope.lifecycleevent
@@ -33,6 +35,7 @@ if [ -x "$(which pypy)" ]; then
 
     virtualenv --clear --no-site-packages -p pypy pypy
 
+    ./pypy/bin/pip install wheel
     ./pypy/bin/pip install coverage
     ./pypy/bin/pip install zope.interface
     ./pypy/bin/pip install zope.lifecycleevent