+++ src/node/behaviors/order.py
@@ -57,3 +57,21 @@ class Order(Behavior):
         if hasattr(self, 'node'):
             if self.node(node.uuid) is not None:
                 raise KeyError('Given node already contained in tree.')
+
+    @override
+    def next_key(self, key):
+        return self.storage.next_key(key)
+
+    @override
+    def prev_key(self, key):
+        return self.storage.prev_key(key)
+
+    @override
+    @property
+    def first_key(self):
+        return self.storage.first_key
+
+    @override
+    @property
+    def last_key(self):
+        return self.storage.last_key